#260 Added code to Enemy script to have necessary base functionality *have not added any code to actually move the enemy or give it AI, but gave it necessary methods and fields to allow this to be done *removed the SetFace methods from the PlayerController and Enemy scripts, and placed the full implementation into the Actor script *created the SetFacingDirection in the Enemy script, allowing the enemy to face different directions while updating it's sprites *added fields to be able to store delegate methods for the attack of the enemy, as well as any abilities they may have, it will differ from enemy to enemy, but some enemies will have just an attack, some will have an attack in the form of an ability, and others will have a regular attack and a non-combat ability *added fields and methods to the Enemy script to allow it to store and update cooldowns for the attack and ability *added code to properly change the facing direction of the Enemy when they are hit

+++ MonsterIsland/Assets/Scripts/MonsterScripts/Monster.cs
@@ -25,7 +25,7 @@ public class Monster : MonoBehaviour {
 
     //used to change the direction the monster is facing, while keeping each
     //arm on the correct side, scaleX is the value to be applied to each part's local scale x
-    public void ChangeDirection(int scaleX)
+    public void ChangeDirection(float scaleX)
     {
         //setting the scale of the monster object (a child of, but different from the player object)
         transform.localScale = new Vector2(scaleX, transform.localScale.y);
